refactor: updating the proto file for multisig
